Add Stack Overflow and Twitter nav links, shrink logos (#26)
These are some small tweaks to the website. I added our Twitter and Stack Overflow in the nav at the top. And I also shrunk the "powered by" company logos a bit to make them fit more nicely (especially on mobile, where there used to be only one per line), set them to open in new tabs, and updated some of the text with links to the community, quick start guide, etc.